+++ packages/react-magma-dom/src/components/Dropdown/DropdownSplitButton.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 
-import { Button, ButtonShape, ButtonStyles } from '../Button';
+import { Button, ButtonShape, ButtonStyles, ButtonVariant } from '../Button';
 import { IconButton } from '../IconButton';
-import { ArrowDropUpIcon, CaretDownIcon } from 'react-magma-icons';
-import { DropdownContext, DropdownDropDirection } from '.';
+import { ArrowDropUpIcon, ArrowDropDownIcon } from 'react-magma-icons';
+import { DropdownContext, DropdownDropDirection } from './Dropdown';
 import { I18nContext } from '../../i18n';
 import { useForkedRef, useGenerateId } from '../../utils';
 import { ThemeContext } from '../../theme/ThemeContext';
@@ -35,7 +35,14 @@ export const DropdownSplitButton = React.forwardRef<
   HTMLButtonElement,
   DropdownSplitButtonProps
 >((props, forwardedRef) => {
-  const { 'aria-label': ariaLabel, children, id, onClick, ...other } = props;
+  const {
+    'aria-label': ariaLabel,
+    children,
+    id,
+    variant = ButtonVariant.solid,
+    onClick,
+    ...other
+  } = props;
 
   const context = React.useContext(DropdownContext);
 
@@ -49,7 +56,7 @@ export const DropdownSplitButton = React.forwardRef<
     context.dropDirection === DropdownDropDirection.up ? (
       <ArrowDropUpIcon size={theme.iconSizes.medium} testId="caretUp" />
     ) : (
-      <CaretDownIcon size={theme.iconSizes.medium} testId="caretDown" />
+      <ArrowDropDownIcon size={theme.iconSizes.medium} testId="caretDown" />
     );
 
   function handleClick(event: React.SyntheticEvent) {
@@ -70,6 +77,7 @@ export const DropdownSplitButton = React.forwardRef<
         onClick={onClick}
         shape={ButtonShape.leftCap}
         style={{ borderRight: 0, marginRight: 0 }}
+        variant={variant}
       >
         {children}
       </Button>
@@ -81,8 +89,12 @@ export const DropdownSplitButton = React.forwardRef<
         icon={buttonIcon}
         onClick={handleClick}
         shape={ButtonShape.rightCap}
-        style={{ marginLeft: 0 }}
+        style={{
+          marginLeft:
+            variant === ButtonVariant.outline ? 0 : theme.spaceScale.spacing01,
+        }}
         ref={ref}
+        variant={variant}
       />
     </>
   );