+++ packages/react-magma-dom/src/components/StyledButton/index.tsx
@@ -17,7 +17,7 @@ import {
   buildFocusColor,
 } from './styles';
 import { ThemeContext } from '../../theme/ThemeContext';
-import { ButtonProps } from '../Button';
+import { ButtonProps, ButtonType } from '../Button';
 
 interface StyledButtonProps extends ButtonProps {
   href?: string;
@@ -109,13 +109,13 @@ export const buttonStyles = props => css`
   }
 
   ${props.iconOnly &&
-    css`
-      display: inline-flex;
-      justify-content: center;
-      line-height: 1;
-      min-width: 0;
-      padding: 0;
-    `}
+  css`
+    display: inline-flex;
+    justify-content: center;
+    line-height: 1;
+    min-width: 0;
+    padding: 0;
+  `}
 `;
 
 export const BaseStyledButton = styled.button`
@@ -125,8 +125,7 @@ export const BaseStyledButton = styled.button`
 export const StyledButton = React.forwardRef<
   HTMLButtonElement,
   StyledButtonProps
->((props, ref) => {
-  const { children, testId, ...other } = props;
+>(({ children, testId, type = ButtonType.button, ...other }, ref) => {
   const theme = React.useContext(ThemeContext);
 
   return (