+++ packages/react-magma-dom/src/index.ts
@@ -78,6 +78,7 @@ export {
   DropdownButton,
   DropdownButtonProps,
 } from './components/Dropdown/DropdownButton';
+export * from './components/Flex';
 export { Form, FormProps } from './components/Form';
 export { FormGroup, FormGroupProps } from './components/FormGroup';
 export { Heading, HeadingProps } from './components/Heading';
@@ -134,6 +135,7 @@ export {
   SkipLinkContent,
   SkipLinkContentProps,
 } from './components/SkipLinkContent';
+export { Spacer, SpacerProps, SpacerAxis } from './components/Spacer';
 export { Spinner, SpinnerProps } from './components/Spinner';
 export {
   Table,