+++ website/react-magma-docs/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.1.9-alpha.0 (2021-03-25)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8 (2021-03-24)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.13 (2021-03-23)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.12 (2021-03-23)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.11 (2021-03-18)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.10 (2021-03-17)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.9 (2021-03-15)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.8 (2021-03-15)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.7 (2021-03-15)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
 ### 2.1.8-alpha.6 (2021-03-12)
 
 **Note:** Version bump only for package react-magma-docs